dev_crew/requirements.txt
crewai
langchain_openai
python-dotenv
litellm